requirements.txt
PyQt5>=5.14
pathlib
imagesize
matplotlib
pillow
pandas